requirements.txt
# FAL Serverless Requirements (CPU Optimized)
# These are needed for both deployment and client

# Core
fal-client>=0.4.0

# For prompt generation and optimization
openai>=1.0.0

# Utilities
python-dotenv
aiohttp

# Enhanced validation and monitoring
pydantic>=2.0.0
orjson>=3.9.0

